refactor: Split RecurrencePage comp into different files
nothing changes
